Separate routes from entry point and modularize components, add new eslint rules,
