+++ .hound.yml
@@ -1,4 +1,6 @@
 eslint:
   enabled: true
   config_file: .eslintrc.json
-  
+
+scss: 
+  enabled: false
