+++ client/index.html
@@ -25,6 +25,7 @@
 			<div id="app"></div>
 			<!-- <script type="text/javascript" src="/static/bundle.js"></script> -->
 			<script type="text/javascript" src="/dist/bundle.js"></script>
+			<!-- <script type="text/javascript" src="/dist/bundle.min.js"></script> -->
 			<!-- Bootstrap.js, Jquery.js, Popper.js -->
 			<script src="https://code.jquery.com/jquery-3.1.1.slim.min.js" integrity="sha384-A7FZj7v+d/sdmMqp/nOQwliLvUsJfDHW+k9Omg/a/EheAdgtzNs3hpfag6Ed950n" crossorigin="anonymous"></script>
 			<script src="https://cdnjs.cloudflare.com/ajax/libs/tether/1.4.0/js/tether.min.js" integrity="sha384-DztdAPBWPRXSA/3eYEEUWrWCy7G5KFbe8fFjk5JAIxUYHKkDx6Qin1DkWx51bBrb" crossorigin="anonymous"></script>
