+++ client/src/store.js
@@ -3,8 +3,9 @@ import { createStore, applyMiddleware,
 } from 'redux';
 import thunk from 'redux-thunk';
 
-// import rootReducer from './rootReducer';
+import rootReducer from './reducers/indexReducer';
 import userReducer from './reducers/userReducer';
+
 
 
 const reducer = combineReducers({
