+++ package.json
@@ -96,7 +96,6 @@
     "style-loader": "^0.19.0",
     "supertest": "^3.0.0",
     "swagger-jsdoc": "^1.9.7",
-    "uglifyjs-webpack-plugin": "^1.1.2",
     "validator": "^9.1.1",
     "webpack": "^3.8.1",
     "webpack-dev-middleware": "^1.12.1",
